Abstract out what actually launches the command
Running a command with a modified environment creates a race condition
where the ENV can get clobbered. Process.spawn and POSIX::Spawn.spawn
get around this by allowing an env hash to be passed in, but they are
not available on 1.8. So, if they're available, use them, and if they're
not, fall back to the backticks.