Interpolation tweaks for not-entire-words
Adds a word-break to the regexp that does interpolations. Now, if you
have :arg1 and :arg11 as interpolations in your command, they will be
replaced properly. However, if you have :arg11 in the line and :arg1 in
the variables hash, the replacement will not be the same as :{arg1}1.
Unsure if this will be an issue, though.