Add a semicolon to the PATH definition
So, this seems trivial, but according to #77, the previous behavior
caused a problem with more complicated commands. Granted, it seems weird
to call commands like `if [thing]; then true; else false; fi` but here
we are. If I add the semicolon, then the shell treats it as a separate
command. This means it modifies the shell env, but this is fine in the
subshell environment we're running the command in. Normally, if you were
running this on your command line, I wouldn't recommend it: as I said,
this would modify your environment instead of modifying the command's
environment. But we should be good here.