Java does not need to run commands with `env`
We added running the commands via `env` for Java because they failed to
run with the modified environment correctly. Using `env` was a kind of
band-aid, I guess. Turns out that the fix to add semicolons obviated the
need to run `env`. Funny thing was that it caused a different problem
when we had both. But since getting rid of it seems to work, we have a
solution.